// portal/src/app/models/signup-req.ts
export interface SignUpRequest {
    firstName: String;
    lastName: String;
    email: String;
    phone: String;
    username: String;
    password: String;

    /* -----------------------------
        private String firstName;
        private String lastName;
        private String username;
        private String email;
        private String password;
        private String phone;
    ------------------------------- */
}